+++ drivers/azure/azureutil/cleanup.go
@@ -1,10 +1,11 @@
 package azureutil
 
 import (
+	"context"
 	"fmt"
 
-	"github.com/Azure/azure-sdk-for-go/arm/compute"
-	"github.com/Azure/azure-sdk-for-go/arm/network"
+	compute "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute"
+	network "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork"
 	"github.com/docker/machine/drivers/azure/logutil"
 	"github.com/docker/machine/libmachine/log"
 )
@@ -60,12 +61,15 @@ type subnetCleanup struct {
 }
 
 func (c *subnetCleanup) Get(a AzureClient) (err error) {
-	c.ref, err = a.subnetsClient().Get(c.rg, c.vnet, c.name, "")
+	subnetsClient := a.subnetsClient()
+	rsp, err := subnetsClient.Get(context.TODO(), c.rg, c.vnet, c.name, &network.SubnetsClientGetOptions{})
+	c.ref = rsp.Subnet
 	return err
 }
 
 func (c *subnetCleanup) Delete(a AzureClient) error {
-	_, err := a.subnetsClient().Delete(c.rg, c.vnet, c.name, nil)
+	subnetsClient := a.subnetsClient()
+	_, err := subnetsClient.BeginDelete(context.TODO(), c.rg, c.vnet, c.name, nil)
 	return err
 }
 
@@ -74,7 +78,7 @@ func (c *subnetCleanup) ResourceType() string { return "Subnet" }
 func (c *subnetCleanup) LogFields() logutil.Fields { return logutil.Fields{"name": c.name} }
 
 func (c *subnetCleanup) HasAttachedResources() bool {
-	return c.ref.Properties.IPConfigurations != nil && len(*c.ref.Properties.IPConfigurations) > 0
+	return c.ref.Properties.IPConfigurations != nil && len(c.ref.Properties.IPConfigurations) > 0
 }
 
 // vnetCleanup manages cleanup of Virtual Network resources.
@@ -84,12 +88,15 @@ type vnetCleanup struct {
 }
 
 func (c *vnetCleanup) Get(a AzureClient) (err error) {
-	c.ref, err = a.virtualNetworksClient().Get(c.rg, c.name, "")
+	virtualNetworksClient := a.virtualNetworksClient()
+	rsp, err := virtualNetworksClient.Get(context.TODO(), c.rg, c.name, &network.VirtualNetworksClientGetOptions{})
+	c.ref = rsp.VirtualNetwork
 	return err
 }
 
 func (c *vnetCleanup) Delete(a AzureClient) error {
-	_, err := a.virtualNetworksClient().Delete(c.rg, c.name, nil)
+	virtualNetworksClient := a.virtualNetworksClient()
+	_, err := virtualNetworksClient.BeginDelete(context.TODO(), c.rg, c.name, nil)
 	return err
 }
 
@@ -98,7 +105,7 @@ func (c *vnetCleanup) ResourceType() string { return "Virtual Network" }
 func (c *vnetCleanup) LogFields() logutil.Fields { return logutil.Fields{"name": c.name} }
 
 func (c *vnetCleanup) HasAttachedResources() bool {
-	return c.ref.Properties.Subnets != nil && len(*c.ref.Properties.Subnets) > 0
+	return c.ref.Properties.Subnets != nil && len(c.ref.Properties.Subnets) > 0
 }
 
 // avSetCleanup manages cleanup of Availability Set resources.
@@ -108,12 +115,15 @@ type avSetCleanup struct {
 }
 
 func (c *avSetCleanup) Get(a AzureClient) (err error) {
-	c.ref, err = a.availabilitySetsClient().Get(c.rg, c.name)
+	availabilitySetsClient := a.availabilitySetsClient()
+	rsp, err := availabilitySetsClient.Get(context.TODO(), c.rg, c.name, &compute.AvailabilitySetsClientGetOptions{})
+	c.ref = rsp.AvailabilitySet
 	return err
 }
 
 func (c *avSetCleanup) Delete(a AzureClient) error {
-	_, err := a.availabilitySetsClient().Delete(c.rg, c.name)
+	availabilitySetsClient := a.availabilitySetsClient()
+	_, err := availabilitySetsClient.Delete(context.TODO(), c.rg, c.name, &compute.AvailabilitySetsClientDeleteOptions{})
 	return err
 }
 
@@ -122,5 +132,5 @@ func (c *avSetCleanup) ResourceType() string { return "Availability Set" }
 func (c *avSetCleanup) LogFields() logutil.Fields { return logutil.Fields{"name": c.name} }
 
 func (c *avSetCleanup) HasAttachedResources() bool {
-	return c.ref.Properties.VirtualMachines != nil && len(*c.ref.Properties.VirtualMachines) > 0
+	return c.ref.Properties.VirtualMachines != nil && len(c.ref.Properties.VirtualMachines) > 0
 }