+++ libmachine/mcndockerclient/docker_versioner.go
@@ -23,7 +23,7 @@ func (dv *defaultDockerVersioner) DockerVersion(host DockerHost) (string, error)
 		return "", fmt.Errorf("Unable to query docker version: %s", err)
 	}
 
-	version, err := client.ServerVersion(context.Background())
+	version, err := client.ServerVersion(context.TODO())
 	if err != nil {
 		return "", fmt.Errorf("Unable to query docker version: %s", err)
 	}